plugin: update `resolveSystemPlugins` default
The commit updates the `resolveSystemPlugins` backend application prop
to `false` since it has the better behavior. The flag was initially set
to `true` not to have breaking behavior changes as part of `v1.19.0`.

The prop is used to not resolve system extension-packs and dependencies
at runtime, and instead resolve them at buildtime.

Signed-off-by: vince-fugnitto <vincent.fugnitto@ericsson.com>